merge from main from official

+++ src/include/duckdb/storage/compression/chimp/algorithm/bit_reader.hpp
@@ -8,7 +8,6 @@
 
 #pragma once
 
-#include "duckdb.h"
 #include "duckdb/common/numeric_utils.hpp"
 
 namespace duckdb {
@@ -150,7 +149,7 @@ public:
 			result = result << 8 | InnerReadByte(i);
 		}
 		result = result << remainder | InnerRead(remainder, bytes);
-		index += (bytes << 3) + remainder;
+		index += static_cast<uint32_t>(bytes << 3) + remainder;
 		return result;
 	}
